mv code to function
